requirements.txt
Flask 
Flask-JWT-Extended
Flask-Smorest
Flask-SQLAlchemy
Flask-Migrate
passlib
marshmallow
python-dotenv
gunicorn
requests 
